backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
requests>=2.31.0
openai==1.3.5
xgboost==2.0.2
pandas>=2.0.0
numpy==1.26.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
geopy==2.4.1
shapely==2.0.2
geoalchemy2==0.14.2
pytest==7.4.3
httpx==0.25.2
loguru==0.7.2
certifi==2024.2.2
torch
faiss-cpu==1.7.4 
scikit-learn==1.3.2 